Fix 32bits issue on literal (#530)

+++ include/boost/simd/literal.hpp
@@ -20,22 +20,22 @@ namespace boost { namespace simd
 
     // Terminal case for char recursion
     template< std::size_t N>
-    constexpr std::uint64_t parse(int,const char (&)[N], nsm::uint64_t<0> const&)
+    constexpr std::uint32_t parse(int,const char (&)[N], nsm::uint32_t<0> const&)
     {
       return 0;
     }
 
     // Extract a char, get its value, multiply by current base, repeat
-    template< std::size_t N, std::size_t I>
-    constexpr std::uint64_t parse(int base, const char (&arr)[N], nsm::uint64_t<I> const&)
+    template< std::size_t N, std::uint32_t I>
+    constexpr std::uint32_t parse(int base, const char (&arr)[N], nsm::uint32_t<I> const&)
     {
-      return to_int(arr[I - 1]) * base + parse<N>(base*10, arr,nsm::uint64_t<I-1>{});
+      return to_int(arr[I - 1]) * base + parse<N>(base*10, arr,nsm::uint32_t<I-1>{});
     }
 
     // Intermediate trampoline so MSVC doesn't cry
-    template<std::size_t N,char... c> constexpr std::uint64_t parse(int base)
+    template<std::size_t N,char... c> constexpr std::uint32_t parse(int base)
     {
-      return parse<N>(base, {c...}, nsm::uint64_t<N>{});
+      return parse<N>(base, {c...}, nsm::uint32_t<N>{});
     }
   }
 
@@ -43,7 +43,7 @@ namespace boost { namespace simd
   {
     // Enables xxx_c to be turned into usable integral constant int_<xxx>
     template <char ...c>
-    constexpr nsm::uint64_t< boost::simd::detail::parse<sizeof...(c),c...>(1)> operator"" _c()
+    constexpr nsm::uint32_t< boost::simd::detail::parse<sizeof...(c),c...>(1)> operator"" _c()
     {
       return {};
     }
